refine wizard refactor planning
